backend/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colour==0.1.5
contextlib2==21.6.0
cryptography==45.0.7
deepdiff==8.6.1
Django==5.2.6
django-cors-headers==4.8.0
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drawsvg==2.4.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
injector==0.22.0
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.3
openpyxl==3.1.5
orderly-set==5.5.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
stix2==3.0.1
stix2-patterns==2.0.0
tabulate==0.9.0
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
xlsxwriter==3.2.5
